LayoutTests/css3/flexbox/flex-item-firstLine-valid.html

<!DOCTYPE html>
<html>
<link href="resources/flexbox.css" rel="stylesheet">
<style>
    div::first-line { line-height: 100px; }
    p::first-line { line-height: 100px;}
</style>

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
<body onload="checkLayout('.flexbox')">
<div id=log></div>
<p>This test flex item should ignore flex container's firstLine pseudo element, but flex item's firstLine should be valid.</p>

<div class="flexbox">
    <p data-expected-height=100>The first item.</p>
    <p data-expected-height=100>The second item.</p>
</div>
</body>
</html>
